Linkify moves in study comments
Closes #3483.

Does not apply to Gamebook and Practice comments.
