+++ ui/chat/src/enhance.ts
@@ -1,3 +1,5 @@
+import { addPlies } from 'common';
+
 export default function(text: string, parseMoves: boolean): string {
   const escaped = window.lichess.escapeHtml(text);
   const linked = autoLink(escaped);
@@ -23,15 +25,4 @@ function userLinkReplace(orig: string, prefix: String, user: string) {
 
 function autoLink(html: string) {
   return html.replace(userPattern, userLinkReplace).replace(linkPattern, linkReplace);
-}
-
-const movePattern = /\b(\d+)\s?(\.+)\s?(?:[o0-]+|[NBRQK]*[a-h]?[1-8]?x?@?[a-h][0-9]=?[NBRQK]?)\+?\#?[!\?=]*/gi;
-function moveReplacer(match: string, turn: number, dots: string) {
-  if (turn < 1 || turn > 200) return match;
-  const ply = turn * 2 - (dots.length > 1 ? 0 : 1);
-  return '<a class="jump" data-ply="' + ply + '">' + match + '</a>';
-}
-
-function addPlies(html: string) {
-  return html.replace(movePattern, moveReplacer);
 }
